cma: Create a tool to create checks against cma block memory  spaces
Introduce a frameork to collect the device's CMA memory blocks, and
functions able to check a given memory space against the blocks to see
if the accessed space given is a valid CMA memory request.

Use this to refuse access to unauthorized addresses given to the mmap
funtions in the mem-exynos and s5p-smem char devices.
Signed-off-by: Curtis Menard <curtis.menard@gmail.com>

Conflicts:

	arch/arm/mach-exynos/mach-midas.c
	arch/arm/mach-exynos/secmem-allocdev.c
	drivers/char/exynos_mem.c